Перегруппировал компоненты в стиле FSD, сделал апи для модулей, вырезал устаревшие модели.
