--- src/components/Main/utils.ts
@@ -1,2 +1,0 @@
-export const getIconPath = (iconName: string, iconType: 'svg' | 'jpg' | 'png' | 'gif' = 'svg') =>
-    `img/animated/${iconName}.${iconType}`;
