--- src/components/Preloader/style.scss
@@ -1,25 +1,0 @@
-.loader-page {
-  position: absolute;
-  z-index: 9000;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: #e6eefc;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-    h1, p{
-      color: #4b4b4b;
-    }
-
-    h1 {
-      font-size: 1.75rem;
-    }
-
-    p {
-      font-size: 1.25rem;
-    }
-}
