+++ src/entities/index.ts
@@ -1,0 +1,2 @@
+export * from './weather';
+export * from './weather-forecast';
