+++ src/entities/weather/model/consts.ts
