+++ src/entities/weather/model/index.ts
@@ -1,0 +1,2 @@
+export * from './weather';
+export * from './consts';
