+++ src/entities/weather/model/weather.ts
