+++ src/entities/weather/ui/ForecastCard/style.scss
