+++ src/entities/weather/ui/index.ts
@@ -1,0 +1,1 @@
+export { ForecastCard } from './ForecastCard';
