--- src/models/index.ts
@@ -1,3 +1,0 @@
-export * from './weather';
-export * from './weather-forecast';
-export * from './open-meteo';
