+++ src/pages/home/index.scss
@@ -1,4 +1,3 @@
-
 .weather-app-main {
   color: white;
   height: 100%;
