+++ src/shared/index.ts
@@ -1,1 +1,5 @@
 export * from './consts';
+export * from './utils';
+export * from './services';
+export * from './interfaces';
+export * from './ui';
