+++ src/shared/ui/CelsiusSign/index.scss
@@ -1,0 +1,4 @@
+span { // TODO: доделать
+  padding-left: 10px;
+  font-size: 50px;
+}
