+++ src/shared/utils/getWeatherIconPath.ts
@@ -1,0 +1,2 @@
+export const getWeatherIconPath = (iconName: string, iconType: 'svg' | 'jpg' | 'png' | 'gif' = 'svg') =>
+    `img/animated/${iconName}.${iconType}`;
