+++ src/shared/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './cn';
+export * from './getWeatherIconPath';
