+++ src/widgets/ErrorBoundary/index.tsx
