+++ src/widgets/Header/index.tsx
@@ -1,9 +1,8 @@
-import React from 'react';
 import './style.scss';
 import {useStore} from 'effector-react';
-import {$currentWeather} from '@models';
 import {Button} from 'primereact/button';
 import {useClassName} from '@utils';
+import { FC } from 'react';
 
 export type HeaderPropType = {
     latitude: number;
@@ -12,13 +11,13 @@ export type HeaderPropType = {
     sideMenuVisible: boolean;
 };
 
-const Header: React.FC<HeaderPropType> = ({sideMenuVisible, btnAction }) => {
+const Header: FC<HeaderPropType> = ({sideMenuVisible, btnAction }) => {
     const cn = useClassName('weather-app-header');
 
-    const { currentWeather, isLoading} = useStore($currentWeather);
+    // const { currentWeather, isLoading} = useStore($currentWeather);
     const currentDate = new Date().toLocaleDateString();
-    const {country} = currentWeather.sys;
-    const city = currentWeather.name;
+    // const {country} = currentWeather.sys;
+    // const city = currentWeather.name;
 
     return (
         <header className={cn()}>
@@ -35,7 +34,7 @@ const Header: React.FC<HeaderPropType> = ({sideMenuVisible, btnAction }) => {
                 }
             </div>
             <div>
-                <p className={cn('position')}>{country} - {city}</p>
+                {/* <p className={cn('position')}>{country} - {city}</p> */}
             </div>
             <div>
                 <p className={cn('current-date')}>{currentDate}</p>
