+++ src/widgets/Loader/style.scss
