+++ src/widgets/Sidebar/style.scss
