+++ src/widgets/index.ts
@@ -1,6 +1,4 @@
-export {default as Main} from './Main';
-export {default as Sidebar} from './Sidebar';
-export {default as Preloader} from './Preloader';
 export {default as Header} from './Header';
 export {default as ErrorBoundary} from './ErrorBoundary';
 export {default as Loader} from './Loader';
+export {default as Sidebar} from './Sidebar';
