+++ tsconfig.paths.json
@@ -1,6 +1,9 @@
 {
   "compilerOptions": {
     "paths": {
+      "@src": [
+        "src"
+      ],
       "@shared": [
         "src/shared"
       ],
@@ -18,6 +21,18 @@
       ],
       "@models": [
         "src/models"
+      ],
+      "@entities": [
+        "src/entities"
+      ],
+      "@pages": [
+        "src/pages"
+      ],
+      "@features": [
+        "src/features"
+      ],
+      "@widgets": [
+        "src/widgets"
       ]
     }
   }
